Add a --version switch for recording file version strings.
In automation usage it's difficult to use versioned filename
like in a traditional release tarball because scripts would
need to parse out and use the versioned directory after
unpacking.

It has therefore become fasionable to include a 'version' key in
manifest files to identifying the specific revision represented
by a hash. This commits adds a command-line switch to set this
when adding a file for convenience.